Amends templating tests
